feat: 封装 NewsList 组件
