+++ src/apis/news/interface.ts
@@ -1,0 +1,10 @@
+export interface INews {
+  title: string;
+  content: string;
+  time: string;
+  link: string;
+}
+
+export interface INewsList {
+  messageList: INews[];
+}
