+++ src/pages/News/index.module.less
@@ -1,0 +1,34 @@
+.newsWrapper {
+  width: 100%;
+  display: flex;
+  justify-content: center;
+
+  .newsContainer {
+    width: 100%;
+    max-width: 800px;
+    padding: 16px;
+    margin: 0 auto;
+  }
+}
+
+.newsItem {
+  width: 100%;
+  box-shadow: 0 6px 16px 0 rgba(0, 0, 0, 0.08);
+  cursor: pointer;
+}
+
+.newsLoading {
+  margin: 1rem 0;
+  text-align: center;
+}
+
+.backTop {
+  height: 40px;
+  width: 40px;
+  line-height: 40px;
+  border-radius: 4px;
+  background-color: #1088e9;
+  color: #fff;
+  text-align: center;
+  font-size: 14px;
+}
